add new features

+++ src/UserFunctions.ts
@@ -1,37 +1,69 @@
 import prisma from './primsaClient'
 import { v4 } from 'uuid'
 import { hashData, compareHash } from '@jabz/security-utils';
-import { getEnvVar } from './utilities';
+import { UserCreationErrors, UserCreationResults, getEnvVar, objectsEqual, passWordCheck, validateEmail } from './utilities';
+import { TwoFactorClientData, createNewAuthDevice } from './2Factor/twofactor';
 
 const userClient = prisma.user;
 
 
 export interface NewUser {
     email: string
+    userName: string
     password: string
+    authDevice: TwoFactorClientData
 }
 
 /**
  * creates a single user
 */
-export async function createSingleUser(newUserData: NewUser, idGenerator?: (newUserData?: NewUser) => string) {
+export async function createSingleUser(newUserData: NewUser, idGenerator?: (newUserData?: NewUser) => string): Promise<UserCreationResults> {
     try {
-        const validator = require("email-validator");
-        if (!validator.validate(newUserData.password)) return false
+        const check = Object.values(newUserData.authDevice).map((e) => Boolean(e)).every((r) => r)
+        if (!check) return { isValid: false, errorMessage: [UserCreationErrors.deviceDataMissing] }
+        const exist = await userClient.findMany({ where: { userName: newUserData.userName } })
+        if (exist.length) return { isValid: false, errorMessage: [UserCreationErrors.userExist] }
         const result = await userClient.create({
             data: {
                 id: idGenerator ? idGenerator(newUserData) : v4(),
                 email: newUserData.email,
+                userName: newUserData.userName,
                 password: await hashData(newUserData.password, 0),
                 createdAt: Date.now(),
+                lastPasswordUpdate: BigInt(Date.now()),
+                authDevice: [JSON.stringify(createNewAuthDevice(newUserData.authDevice, true))],
             }
         })
-        return Boolean(result);
+        return { isValid: Boolean(result) };
+    } catch (error) {
+        console.error(error);
+        return { isValid: false, errorMessage: ["There seems to be a server error"] }
+    }
+
+}
+
+export async function verifyPassword(userID: string, authDevice: TwoFactorClientData, password: string) {
+    try {
+        const user = await userClient.findFirst({
+            where: { OR: [{ id: userID }, { email: userID }, { userName: userID }] },
+            select: {
+                id: false,
+                email: false,
+                password: true,
+                lastPasswordUpdate: false,
+                createdAt: false,
+                authDevice: true,
+            }
+        });
+
+        if (objectsEqual(authDevice, JSON.parse(user?.authDevice as string ?? '{}'))) return undefined;
+        else if (user) return compareHash(password, 0, user?.password);
+        else return null;
+
     } catch (error) {
         console.error(error)
         return null
     }
-
 }
 
 /** create multiple new users*/
@@ -42,11 +74,14 @@ export async function createUsers(newUserData: NewUser[], idGenerator?: (newUser
         const newUser = validUsers.map(async (e) => ({
             id: idGenerator ? idGenerator(e) : v4(),
             email: e.email,
+            userName: e.userName,
             password: await hashData(e.password, getEnvVar()),
             createdAt: Date.now(),
+            authDevice: JSON.stringify(e.authDevice),
+            lastPasswordUpdate: BigInt(Date.now()),
         }))
         const data = await Promise.all(newUser)
-        const result = await userClient.createMany({ data })
+        const result = await userClient.createMany({ data });
         return Boolean(result)
     } catch (error) {
         console.error(error)
@@ -54,17 +89,6 @@ export async function createUsers(newUserData: NewUser[], idGenerator?: (newUser
     }
 }
 
-/** confirm username password*/
-export async function confirmUserNameAndPassword(userEmail: string, userRawPassword: string) {
-    try {
-        const result = await userClient.findFirst({ where: { email: userEmail } })
-        if (result?.password) return await compareHash(userRawPassword, 0, result.password)
-        else return false
-    } catch (error) {
-        console.log(error)
-        return null
-    }
-}
 
 export interface NewUserUpdate {
     newPassword: string | undefined
@@ -72,13 +96,13 @@ export interface NewUserUpdate {
 
 }
 
-export async function tryUpdateUser(userEmail: string, userOldRawPassword: string, newData: NewUserUpdate) {
+export async function tryUpdateUser(userID: string, authDevice: TwoFactorClientData, userOldRawPassword: string, newData: NewUserUpdate) {
 
     try {
-        const found = await confirmUserNameAndPassword(userEmail, userOldRawPassword)
+        const found = await verifyPassword(userID, authDevice, userOldRawPassword)
         if (found) {
-            const result = await userClient.update({
-                where: { email: userEmail },
+            const result = await userClient.updateMany({
+                where: { OR: [{ id: userID }, { email: userID }, { userName: userID }] },
                 data: {
                     email: newData.newEmail,
                     password: newData.newPassword,
@@ -98,7 +122,7 @@ export async function tryUpdateUser(userEmail: string, userOldRawPassword: strin
 export async function findUser(userID: string[]) {
     try {
         return await userClient.findMany({
-            where: { OR: [{ id: { in: userID } }, { email: { in: userID } }] },
+            where: { OR: [{ id: { in: userID } }, { email: { in: userID } }, { userName: { in: userID } }] },
             select: { id: true, email: true, createdAt: true }
         })
     } catch (error) {
@@ -111,7 +135,7 @@ export async function findUser(userID: string[]) {
 /** delete users by ID*/
 export async function deleteUsersByID(userID: string[]) {
     try {
-        const result = await userClient.deleteMany({ where: { id: { in: userID } } })
+        const result = await userClient.deleteMany({ where: { OR: [{ id: { in: userID } }, { email: { in: userID } }, { userName: { in: userID } }] } })
         return Boolean(result.count)
     } catch (error) {
         return null;
